Support multiple list metadata loaders
