+++ src/Sulu/Bundle/AdminBundle/Metadata/ListMetadata/ListMetadataLoaderInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) Sulu GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\AdminBundle\Metadata\ListMetadata;
+
+interface ListMetadataLoaderInterface
+{
+    public function getMetadata(string $key, string $locale, array $metadataOptions);
+}
