+++ src/Sulu/Bundle/AdminBundle/Metadata/ListMetadata/ListMetadataProvider.php
@@ -13,47 +13,31 @@ namespace Sulu\Bundle\AdminBundle\Metadata\ListMetadata;
 
 use Sulu\Bundle\AdminBundle\Exception\MetadataNotFoundException;
 use Sulu\Bundle\AdminBundle\Metadata\MetadataProviderInterface;
-use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactoryInterface;
-use Symfony\Component\Translation\TranslatorInterface;
 
 class ListMetadataProvider implements MetadataProviderInterface
 {
     /**
-     * @var FieldDescriptorFactoryInterface
+     * @var ListMetadataLoaderInterface[]
      */
-    private $fieldDescriptorFactory;
+    private $listMetadataLoaders;
 
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
-
-    public function __construct(
-        FieldDescriptorFactoryInterface $fieldDescriptorFactory,
-        TranslatorInterface $translator
-    ) {
-        $this->fieldDescriptorFactory = $fieldDescriptorFactory;
-        $this->translator = $translator;
+    public function __construct(iterable $listMetadataLoaders)
+    {
+        $this->listMetadataLoaders = $listMetadataLoaders;
     }
 
-    public function getMetadata(string $key, string $locale, array $metadataOptions)
+    public function getMetadata(string $key, string $locale, array $metadataOptions): ?ListMetadata
     {
-        $fieldDescriptors = $this->fieldDescriptorFactory->getFieldDescriptors($key);
-
-        if (!$fieldDescriptors) {
-            throw new MetadataNotFoundException('list', $key);
+        $list = null;
+        foreach ($this->listMetadataLoaders as $listMetadataLoader) {
+            $list = $listMetadataLoader->getMetadata($key, $locale, $metadataOptions);
+            if ($list) {
+                break;
+            }
         }
 
-        $list = new ListMetadata();
-        foreach ($fieldDescriptors as $fieldDescriptor) {
-            $field = new FieldMetadata($fieldDescriptor->getName());
-
-            $field->setLabel($this->translator->trans($fieldDescriptor->getTranslation(), [], 'admin', $locale));
-            $field->setType($fieldDescriptor->getType());
-            $field->setVisibility($fieldDescriptor->getVisibility());
-            $field->setSortable($fieldDescriptor->getSortable());
-
-            $list->addField($field);
+        if (!$list) {
+            throw new MetadataNotFoundException('list', $key);
         }
 
         return $list;
