+++ src/Sulu/Bundle/AdminBundle/Metadata/ListMetadata/XmlListMetadataLoader.php
@@ -1,0 +1,59 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) Sulu GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\AdminBundle\Metadata\ListMetadata;
+
+use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactoryInterface;
+use Symfony\Component\Translation\TranslatorInterface;
+
+class XmlListMetadataLoader implements ListMetadataLoaderInterface
+{
+    /**
+     * @var FieldDescriptorFactoryInterface
+     */
+    private $fieldDescriptorFactory;
+
+    /**
+     * @var TranslatorInterface
+     */
+    private $translator;
+
+    public function __construct(
+        FieldDescriptorFactoryInterface $fieldDescriptorFactory,
+        TranslatorInterface $translator
+    ) {
+        $this->fieldDescriptorFactory = $fieldDescriptorFactory;
+        $this->translator = $translator;
+    }
+
+    public function getMetadata(string $key, string $locale, array $metadataOptions): ?ListMetadata
+    {
+        $fieldDescriptors = $this->fieldDescriptorFactory->getFieldDescriptors($key);
+
+        if (!$fieldDescriptors) {
+            return null;
+        }
+
+        $list = new ListMetadata();
+        foreach ($fieldDescriptors as $fieldDescriptor) {
+            $field = new FieldMetadata($fieldDescriptor->getName());
+
+            $field->setLabel($this->translator->trans($fieldDescriptor->getTranslation(), [], 'admin', $locale));
+            $field->setType($fieldDescriptor->getType());
+            $field->setVisibility($fieldDescriptor->getVisibility());
+            $field->setSortable($fieldDescriptor->getSortable());
+
+            $list->addField($field);
+        }
+
+        return $list;
+    }
+}
